SRDiff_main/pyproject.toml
[project]
name = "docs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"tqdm",
	"matplotlib",
	"nibabel>=5.3.3",
	"numpy>=2.4.2",
	"scikit-learn>=1.8.0",
	"scikit-image>=0.26.0",
	"opencv-python>=4.13.0.92",
	"pyyaml>=6.0.3",
	"einops>=0.8.2",
	"ipykernel>=7.2.0",
	"requests>=2.32.5",
	"pillow>=12.1.1",
	"pandas>=3.0.1",
	"tensorboard>=2.20.0",
	"lpips==0.1.3",
]
